feat: add Bash 5.3.15 and Readline compatibility
Implement .cherubrc startup, programmable completion, UTF-8 interactive editing, and the remaining job, history, expansion, and timing behavior covered by the reference suites.

Add the Readline and History C ABI, pinned GNU and Brush parity gates, practical examples, and Linux CI.

+++ crates/builtins/src/ulimit.rs
@@ -287,7 +287,7 @@ fn print_one(spec: LimitSpec, value: libc::rlim_t, print_description: bool) {
     if value == libc::RLIM_INFINITY {
         println!("unlimited");
     } else {
-        println!("{}", (value as u64) / spec.factor);
+        println!("{}", value / spec.factor);
     }
 }
